# backend/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for IELTS Vocabulary Learning Platform",
        "title": "IELTS Learning API",
        "contact": {},
        "version": "v1"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/flashcards/review": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Save a flashcard rating and update authenticated user vocabulary progress using MVP spaced repetition rules.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flashcards"
                ],
                "summary": "Save flashcard review",
                "parameters": [
                    {
                        "description": "Flashcard review request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/flashcard.ReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/flashcard.ReviewResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/lessons/{lessonId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return lesson metadata, topic and band context, vocabulary list, and authenticated user progress.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lessons"
                ],
                "summary": "Get lesson detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson ID",
                        "name": "lessonId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/lesson.DetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/lessons/{lessonId}/flashcards": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return flashcards for a lesson with vocabulary content and authenticated user vocabulary progress.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flashcards"
                ],
                "summary": "Get lesson flashcards",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson ID",
                        "name": "lessonId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/flashcard.LessonFlashcardsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/lessons/{lessonId}/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create or update authenticated user lesson progress and mark the lesson as in progress when allowed.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Lessons"
                ],
                "summary": "Start lesson",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lesson ID",
                        "name": "lessonId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/lesson.StartResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/reviews/due": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Return due review flashcards for the authenticated user where next review time is due.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reviews"
                ],
                "summary": "Get due reviews",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Maximum number of due cards, max 100",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by topic ID",
                        "name": "topicId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/flashcard.DueReviewsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "flashcard.DueReviewsResponse": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/flashcard.FlashcardResponse"
                    }
                },
                "count": {
                    "type": "integer"
                }
            }
        },
...
        "flashcard.LessonFlashcardsResponse": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/flashcard.FlashcardResponse"
                    }
                },
                "lesson": {
                    "$ref": "#/definitions/flashcard.LessonResponse"
                },
                "progress": {
                    "$ref": "#/definitions/flashcard.SessionProgress"
                }
            }
        },
        "flashcard.LessonResponse": {
            "type": "object",
            "properties": {
                "bandLabel": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "topicId": {
                    "type": "integer"
                },
                "topicTitle": {
                    "type": "string"
                }
            }
        },
        "flashcard.ReviewRequest": {
            "type": "object",
            "properties": {
                "rating": {
                    "$ref": "#/definitions/models.FlashcardRating"
                },
                "vocabularyId": {
                    "type": "integer"
                }
            }
        },
        "flashcard.ReviewResponse": {
            "type": "object",
            "properties": {
                "correctCount": {
                    "type": "integer"
                },
                "lastRating": {
                    "$ref": "#/definitions/models.FlashcardRating"
                },
                "lastReviewedAt": {
                    "type": "string"
                },
                "masteryScore": {
                    "type": "integer"
                },
                "nextReviewAt": {
                    "type": "string"
                },
                "reviewCount": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/models.VocabularyStatus"
                },
                "totalXp": {
                    "type": "integer"
                },
                "vocabularyId": {
                    "type": "integer"
                },
                "wrongCount": {
                    "type": "integer"
                },
                "xpAwarded": {
                    "type": "integer"
                }
            }
        },
        "flashcard.SessionProgress": {
            "type": "object",
            "properties": {
                "done": {
                    "type": "integer"
                },
                "remaining": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "lesson.BandLevelResponse": {
            "type": "object",
            "properties": {
                "bandScore": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "lesson.DetailResponse": {
            "type": "object",
            "properties": {
                "bandLevel": {
                    "$ref": "#/definitions/lesson.BandLevelResponse"
                },
                "lesson": {
                    "$ref": "#/definitions/lesson.LessonResponse"
                },
                "progress": {
                    "$ref": "#/definitions/lesson.ProgressResponse"
                },
                "topic": {
                    "$ref": "#/definitions/lesson.TopicResponse"
                },
                "vocabularies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lesson.VocabularyResponse"
                    }
                }
            }
        },
        "lesson.LessonResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "estimatedMinutes": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lockedReason": {
                    "type": "string"
                },
                "orderIndex": {
                    "type": "integer"
                },
                "progressPercentage": {
                    "type": "integer"
                },
                "requiredScore": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.LessonStatus"
                },
                "title": {
                    "type": "string"
                },
                "wordCount": {
                    "type": "integer"
                },
                "xpReward": {
                    "type": "integer"
                }
            }
        },
        "lesson.ProgressResponse": {
            "type": "object",
            "properties": {
                "bestScore": {
                    "type": "integer"
                },
                "bestXp": {
                    "type": "integer"
                },
                "completedAt": {
                    "type": "string"
                },
                "lastStudiedAt": {
                    "type": "string"
                },
                "progressPercentage": {
                    "type": "integer"
                },
                "score": {
                    "type": "integer"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.LessonStatus"
                },
                "totalWords": {
                    "type": "integer"
                },
                "wordsLearned": {
                    "type": "integer"
                }
            }
        },
        "lesson.StartResponse": {
            "type": "object",
            "properties": {
                "lastStudiedAt": {
                    "type": "string"
                },
                "lessonId": {
                    "type": "integer"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.LessonStatus"
                }
            }
        },
        "lesson.TopicResponse": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "emoji": {
                    "type": "string"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "slug": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "lesson.VocabularyResponse": {
            "type": "object",
            "properties": {
                "audioUrl": {
                    "type": "string"
                },
                "correctCount": {
                    "type": "integer"
                },
                "difficulty": {
                    "$ref": "#/definitions/models.DifficultyLevel"
                },
                "exampleSentence": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ipa": {
                    "type": "string"
                },
                "lastReviewedAt": {
                    "type": "string"
                },
                "learned": {
                    "type": "boolean"
                },
                "learnedAt": {
                    "type": "string"
                },
                "meaningEn": {
                    "type": "string"
                },
                "meaningVi": {
                    "type": "string"
                },
                "nextReviewAt": {
                    "type": "string"
                },
                "partOfSpeech": {
                    "type": "string"
                },
                "reviewCount": {
                    "type": "integer"
                },
                "shortDefinition": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.VocabularyStatus"
                },
                "targetBand": {
                    "type": "number"
                },
                "word": {
                    "type": "string"
                },
                "wrongCount": {
                    "type": "integer"
                }
            }
        },
...
        "models.FlashcardRating": {
            "type": "string",
            "enum": [
                "AGAIN",
                "HARD",
                "GOOD",
                "EASY"
            ],
            "x-enum-varnames": [
                "FlashcardRatingAgain",
                "FlashcardRatingHard",
                "FlashcardRatingGood",
                "FlashcardRatingEasy"
            ]
        },
        "models.LessonStatus": {
            "type": "string",
            "enum": [
                "LOCKED",
                "UNLOCKED",
                "IN_PROGRESS",
                "COMPLETED"
            ],
            "x-enum-varnames": [
                "LessonStatusLocked",
                "LessonStatusUnlocked",
                "LessonStatusInProgress",
                "LessonStatusCompleted"
            ]
        },
...
        "models.VocabularyStatus": {
            "type": "string",
            "enum": [
                "NEW",
                "LEARNING",
                "REVIEW",
                "MASTERED"
            ],
            "x-enum-varnames": [
                "VocabularyStatusNew",
                "VocabularyStatusLearning",
                "VocabularyStatusReview",
                "VocabularyStatusMastered"
            ]
        },
        "response.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.APIError"
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by a JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}